requirements.txt
# USI-PRO Multi-Agent PDF Anonymizer
# Install with: pip install -r requirements.txt
# Also requires: brew install poppler tesseract

anthropic>=0.39.0
pypdf2>=3.0.0
pdf2image>=1.16.0
pillow>=10.0.0
pytesseract>=0.3.10
pyyaml>=6.0
numpy>=1.24.0
